llm_env/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.29.0
numpy==1.26.0
watchdog==6.0.0
typing-extensions==4.13.2 